+++ src-docs/src/views/provider/provider_example.js
@@ -17,7 +17,6 @@ import GlobalStyles from './provider_styles';
 
 export const ProviderExample = {
   title: 'Provider',
-  isNew: true,
   intro: (
     <EuiText>
       <p>