+++ src-docs/src/views/theme/customizing/_sass.tsx
@@ -1,28 +1,7 @@
-import React, { useContext } from 'react';
-import { EuiSpacer, EuiCallOut, EuiText, EuiLink } from '../../../../../src';
+import React from 'react';
+import { EuiSpacer, EuiText } from '../../../../../src';
 
-// @ts-ignore Importing from JS
-import { ThemeContext } from '../../../components/with_theme';
-import { CustomizeTokens } from '../../guidelines/getting_started/_customize_tokens';
-
-export const SassAlert = () => {
-  const themeContext = useContext(ThemeContext);
-  return themeContext.theme.includes('legacy') ? (
-    <></>
-  ) : (
-    <EuiCallOut color="warning">
-      <p>
-        EUI is transitioning to a CSS-in-JS solution for theming and so this
-        method of overriding the theme will only affect components that are
-        still Sass-based. To track this effort, subscribe to the{' '}
-        <EuiLink href="https://github.com/elastic/eui/issues/3912">
-          Meta ticket
-        </EuiLink>
-        .
-      </p>
-    </EuiCallOut>
-  );
-};
+import { CustomizeTokens } from '../../guidelines/getting_started/_customizing';
 
 export default () => {
   return (